requirements.txt
aiohttp==3.6.2
alpha-vantage==2.1.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
discord==1.0.1
discord.py==1.3.2
idna==2.9
idna-ssl==1.1.0
kiwisolver==1.1.0
matplotlib==3.2.0
mplfinance==0.12.3a3
multidict==4.7.5
numpy==1.18.1
pandas==1.0.2
pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
